版本升级1.2.6

+++ src/main/java/com/qiwenshare/ufop/operation/download/domain/Range.java
@@ -8,6 +8,6 @@ import lombok.Data;
  */
 @Data
 public class Range {
-    private int start;
+    private long start;
     private int length;
 }
